+++ src/aux-common/bots/test/BotCalculationContextTests.ts
@@ -875,6 +875,7 @@ export function botCalculationContextTests(
             ['codeButton'],
             ['codeHint'],
             ['spherePortal'],
+            ['light'],
         ];
         const tagCases = ['auxForm', 'form'];